requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
amqp==5.1.1
aniso8601==7.0.0
anyio==3.6.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.18.21
botocore==1.21.21
brotlipy==0.7.0
celery==5.2.3
certifi==2022.12.7
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.11.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-34_1l0p9/conda-package-handling_1647948700698/work
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==3.4.7
cytoolz==0.11.0
Deprecated==1.2.12
discord-webhook==0.14.0
discord.py==1.7.3
Django==4.1.4
django-admin-autocomplete-filter==0.7.1
django-cache-fallback==0.3.0
django-cacheops==6.1
django-cors-headers==3.8.0
django-debug-toolbar==4.0.0
django-extensions==3.1.5
django-filter==22.1
django-health-check==3.17.0
django-oauth-toolkit==1.5.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-framework-mongoengine==3.4.1
django-storages==1.12
djangoql==0.17.1
djangorestframework==3.14.0
drf-yasg==1.21.4
duckduckgo-images-api==1.0.0
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
exceptiongroup==1.1.0
fabric==2.7.0
fastapi==0.79.0
funcy==1.17
gevent==21.12.0
gi-scraper==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.2.1
httptools==0.4.0
idna==2.10
importlib-metadata==4.6.4
inflection==0.5.1
invoke==1.7.1
ipfs-api==0.2.3
ipfshttpclient==0.7.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
jwcrypto==1.0
kombu==5.2.4
lru-dict==1.1.7
MarkupSafe==2.1.1
mongoengine==0.23.1
multiaddr==0.0.9
multicall @ git+https://github.com/banteg/multicall.py.git@ccfb614bd5bcc90fdf2a2a4dc8cb12ae114f98c5
multidict==5.1.0
netaddr==0.8.0
numpy==1.24.1
oauthlib==3.1.1
openai==0.25.0
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
parallel-ssh==2.10.0
paramiko==2.10.4
parsimonious==0.8.1
pathlib2==2.3.7.post1
pbr==5.6.0
phonenumbers==8.13.10
Pillow==9.0.1
ply==3.11
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.17.3
pssh==2.3.1
psutil==5.8.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
pycosat==0.6.3
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.9.2
pymongo==3.12.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-memcached==1.59
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.25.1
rlp==2.0.1
ruamel-yaml-conda @ file:///Users/ktietz/demo/mc3/conda-bld/ruamel_yaml_1629464769899/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==1.6.1
s3transfer==0.5.0
selenium==3.141.0
sentry-sdk==1.3.1
singledispatch==3.7.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
ssh-python==0.10.0
ssh2-python==0.27.0
starlette==0.19.1
tendo==0.2.15
termcolor==1.1.0
text-unidecode==1.3
toolz==0.11.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2022.7.0.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.18.2
uvloop==0.16.0
varint==1.0.2
vine==5.0.0
watchfiles==0.16.1
wcwidth==0.2.5
web3==5.23.0
webdriver-manager==3.5.2
websockets==10.3
whitenoise==6.4.0
wrapt==1.12.1
wsproto==1.2.0
yarl==1.6.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
